requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
crispy-tailwind==0.5.0
dj-database-url==0.5.0
Django==4.0.4
django-crispy-forms==1.14.0
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.3
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
razorpay==1.3.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.1.0
